Switch to milliseconds for profiling data on poll Add calls to start_step for all computations

+++ geocruncher/profiler/progress.py
@@ -10,9 +10,7 @@ class ProgressRecorder:
     def set_current_task(self, task: Task) -> None:
         self.task = task
 
-    def set_progress(
-        self, current_step: str, start_time: int, total_time: float
-    ) -> None:
+    def set_progress(self, current_step: str, start_time: int, total_time: int) -> None:
         if self.task is None:
             return
         state = AsyncResult(self.task.request.id).state
